added limit switches, unsure of how they will be implemented
