+++ src/main/java/frc/robot/RobotMap.java
@@ -6,8 +6,6 @@ public class RobotMap {
   public final static int            OI_XBOX                              = 0;
   public final static int            OI_BBOARD                            = 0;
   // Climb Constants
-  public final static int            DIO_CLIMB_REAR_LIMIT                 = 0;
-  public final static int            DIO_CLIMB_FRONT_LIMIT                = 0;
   public final static int            PWM_ID_SPARK_WHEELS                  = 0;
   public final static int            PCM_ID_DSOLENOID_CLIMB_FRONT_DEPLOY  = 0;
   public final static int            PCM_ID_DSOLENOID_CLIMB_FRONT_RETRACT = 0;
@@ -15,4 +13,6 @@ public class RobotMap {
   public final static int            PCM_ID_DSOLENOID_CLIMB_REAR_RETRACT  = 0;
   public final static int            ULTRASONIC_SENSOR_FRONT              = 0;
   public final static int            ULTRASONIC_SENSOR_MIDDLE             = 0;
+  public final static int            DIO_CLIMB_REAR_LIMIT                 = 0;
+  public final static int            DIO_CLIMB_FRONT_LIMIT                = 0;
 }
