+++ src/main/java/frc/robot/commands/climb/climbcg/DeployC.java
@@ -1,4 +1,4 @@
-package frc.robot.commands.climbcg;
+package frc.robot.commands.climb.climbcg;
 
 import edu.wpi.first.wpilibj.command.Command;
 import frc.robot.Robot;
